requirements.txt
httpx>=0.27
tenacity>=8.2
python-dotenv>=1.0
pyarrow>=15.0
polars>=1.0
duckdb>=1.0
dropbox>=11.36
